Altered Routes, Completed login/reg
