+++ src/App.js
@@ -6,7 +6,6 @@ import {
   Link
 } from 'react-router-dom';
 import LandingPage from './pages/LandingPage';
-import LoginRegisterPage from './pages/LoginRegisterPage';
 import RecipeListPage from './pages/RecipeListPage';
 import SavedRecipesPage from './pages/SavedRecipesPage';
 import AboutUsPage from './pages/AboutUsPage';
@@ -15,6 +14,8 @@ import SavedRecipesDetail from './pages/SavedRecipesDetail';
 
 import Error from './components/Error';
 import Header from './components/Header';
+import LoginForm from './components/LoginForm';
+import RegisterForm from './components/RegisterForm';
 
 function App() {
   return (
@@ -23,10 +24,12 @@ function App() {
 
       <Routes>
         <Route path='/' element={<LandingPage />} />
-        <Route path='login' element={<LoginRegisterPage />} />
+        <Route path='login' element={<LoginForm />} />
         <Route path='about' element={<AboutUsPage />} />
         <Route path='saved-recipes' element={<SavedRecipesPage />} />
         <Route path='recipes' element={<RecipeListPage />} />
+
+        <Route path='register' element={<RegisterForm />}/>
 
         {/* for testing only */}
         <Route path='detail' element={<SavedRecipesDetail />} />
